backend/requirements.txt
flask
flask-cors
flask-pymongo
flask-bcrypt
flask-jwt-extended
python-dotenv
mysql-connector-python
pymongo
sqlparse
google-generativeai
bson
gunicorn